requirements.txt
pandas
numpy
plotly
scikit-learn
python-dotenv
requests
networkx